backend/requirements.txt
asgiref==3.11.1
django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
gunicorn==23.0.0
Pillow==11.3.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
